correction based on feedback

+++ rover_video/src/aruco_detection/image_capture.hpp
@@ -2,6 +2,7 @@
 #define IMAGE_CAPTURE_HPP
 
 #include "rovus_lib/timer.hpp"
+#include "rovus_lib/ip_pinging.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include <opencv2/aruco.hpp>
 #include <opencv2/opencv.hpp>
@@ -10,8 +11,8 @@
 class ImageCapture
 {
     static constexpr uint64_t DELAY_CAMERA_PINNING_RETRY_MS = 5'000UL;
-    static constexpr uint64_t TIMEOUT_CAMERA_PINNING__MS = 500UL;
-    static constexpr uint16_t CAM_NETWORK_PORT = 554U;
+    static constexpr size_t TIMEOUT_CAMERA_PINNING__MS = 500UL;
+    static constexpr size_t CAM_NETWORK_PORT = 554U;
     static constexpr const char* PIPELINE = " latency=0 drop-on-latency=true protocols=tcp ! "
                                             "decodebin ! "
                                             "videorate max-rate=5 ! "
@@ -28,7 +29,7 @@ class ImageCapture
     cv::Mat getErrorFrame(void);
     bool initCam(void);
     bool isValid(void) const;
-    bool isCameraReachable(const std::string& url_, int port_, int timeoutMs_);
+    bool isCameraReachable(const std::string& url_, size_t port_, size_t timeoutMs_);
 
   private:
     bool _isValid;
